URIs should not be hardcoded
Hard coding a URI makes it difficult to test a program: path literals
are not always portable across operating systems, a given absolute path
may not exist on a specific test environment, a specified Internet URL
may not be available when executing the tests, production environment
filesystems usually differ from the development environment, ...etc. For
 all those reasons, a URI should never be hard coded. Instead, it should
  be replaced by customizable parameter.

Further even if the elements of a URI are obtained dynamically,
portability can still be limited if the path-delimiters are hard-coded.

This rule raises an issue when URI's or path delimiters are hard coded.

//based on sonar (copy pasted message)